requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
cachetools==3.1.1
certifi==2019.9.11
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
cssselect==1.2.0
cssselect2==0.7.0
dj-database-url==0.5.0
Django==3.2.18
django-concurrency==2.2
django-crispy-forms==1.14.0
django-git==0.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-jsignature==0.11
django-libs==2.0.2
django-material==1.11.3
django-optimistic-lock==1.0.0
django-phone-field==1.8.1
django-render-block==0.6
django-selectable==1.3.0
django-templated-email==2.3.0
django-tinymce4-lite==1.7.5
django-widget-tweaks==1.4.3
gitdb2==2.0.5
GitPython==3.0.2
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
gunicorn==20.0.4
html5lib==1.1
httplib2==0.22.0
idna==2.8
jsmin==3.0.0
lxml==4.9.2
oauthlib==3.1.0
oscrypto==1.3.0
Pillow==9.5.0
protobuf==4.22.3
psycopg2==2.9.6
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.21
Pygments==2.4.2
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
pyparsing==3.0.9
pypdf==3.8.1
PyPDF2==3.0.1
pypng==0.20220715.0
pyquery==2.0.0
python-bidi==0.4.2
python-dateutil==2.7.5
python-dotenv==0.15.0
pytz==2018.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.0
six==1.12.0
smmap2==2.0.5
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
Unidecode==1.0.23
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.25.5
webencodings==0.5.1
whitenoise==5.1.0
xhtml2pdf==0.2.10
